+++ linkerd/core/src/main/scala/io/buoyant/linkerd/Linker.scala
@@ -1,17 +1,20 @@
 package io.buoyant.linkerd
 
+import com.twitter.finagle.{Service, http}
 import com.twitter.finagle.buoyant.DstBindingFactory
 import com.twitter.finagle.naming.NameInterpreter
 import com.twitter.finagle.{param, Path, Namer, Stack}
-import com.twitter.finagle.stats.{BroadcastStatsReceiver, DefaultStatsReceiver, NullStatsReceiver}
-import com.twitter.finagle.tracing.{debugTrace => fDebugTrace, NullTracer, DefaultTracer, BroadcastTracer, Tracer}
+import com.twitter.finagle.stats.{BroadcastStatsReceiver, LoadedStatsReceiver}
+import com.twitter.finagle.tracing.{DefaultTracer, BroadcastTracer, Tracer}
 import com.twitter.finagle.util.LoadService
 import com.twitter.logging.Logger
-import io.buoyant.admin.AdminConfig
+import io.buoyant.admin.{AdminConfig, Admin}
 import io.buoyant.config._
+import io.buoyant.config.types.Port
 import io.buoyant.namer.Param.Namers
 import io.buoyant.namer._
-import io.buoyant.telemetry.{TelemeterInitializer, TelemeterConfig, Telemeter}
+import io.buoyant.telemetry._
+import scala.util.control.NoStackTrace
 
 /**
  * Represents the total configuration of a Linkerd process.
@@ -19,7 +22,7 @@ import io.buoyant.telemetry.{TelemeterInitializer, TelemeterConfig, Telemeter}
 trait Linker {
   def routers: Seq[Router]
   def namers: Seq[(Path, Namer)]
-  def admin: AdminConfig
+  def admin: Admin
   def tracer: Tracer
   def telemeters: Seq[Telemeter]
   def configured[T: Stack.Param](t: T): Linker
@@ -28,10 +31,14 @@ trait Linker {
 object Linker {
   private[this] val log = Logger()
 
+  private[this] val DefaultAdminPort = Port(9990)
+  private[this] val DefaultAdminConfig = AdminConfig(Some(DefaultAdminPort))
+
   private[linkerd] case class Initializers(
     protocol: Seq[ProtocolInitializer] = Nil,
     namer: Seq[NamerInitializer] = Nil,
     interpreter: Seq[InterpreterInitializer] = Nil,
+    transformer: Seq[TransformerInitializer] = Nil,
     tlsClient: Seq[TlsClientInitializer] = Nil,
     tracer: Seq[TracerInitializer] = Nil,
     identifier: Seq[IdentifierInitializer] = Nil,
@@ -40,7 +47,7 @@ object Linker {
     announcer: Seq[AnnouncerInitializer] = Nil
   ) {
     def iter: Iterable[Seq[ConfigInitializer]] =
-      Seq(protocol, namer, interpreter, tlsClient, tracer, identifier, classifier, telemetry, announcer)
+      Seq(protocol, namer, interpreter, tlsClient, tracer, identifier, transformer, classifier, telemetry, announcer)
 
     def all: Seq[ConfigInitializer] = iter.flatten.toSeq
 
@@ -55,6 +62,7 @@ object Linker {
     LoadService[ProtocolInitializer],
     LoadService[NamerInitializer],
     LoadService[InterpreterInitializer] :+ DefaultInterpreterInitializer,
+    LoadService[TransformerInitializer],
     LoadService[TlsClientInitializer],
     LoadService[TracerInitializer],
     LoadService[IdentifierInitializer],
@@ -84,88 +92,83 @@ object Linker {
     telemetry: Option[Seq[TelemeterConfig]],
     admin: Option[AdminConfig]
   ) {
-    def mk(): Linker = {
+
+    def mk(defaultTelemeter: Telemeter = NullTelemeter): Linker = {
       // At least one router must be specified
       if (routers.isEmpty) throw NoRoutersSpecified
 
-      val telemeters = telemetry.map(_.map(_.mk(Stack.Params.empty)))
+      val telemeters = telemetry match {
+        case None => Seq(defaultTelemeter)
+        case Some(telemeters) => telemeters.map(_.mk(Stack.Params.empty))
+      }
 
-      // Telemeters may provide StatsReceivers.  Note that if all
-      // telemeters provide implementations that do not use the
-      // default Metrics registry, linker stats may be missing from
-      // /admin/metrics.json
-      val stats =
-        telemeters.getOrElse(Nil).collect { case t if !t.stats.isNull => t.stats } match {
-          case Nil =>
-            log.info(s"Using default stats receiver")
-            DefaultStatsReceiver
-          case receivers =>
-            for (r <- receivers) log.info(s"Using stats receiver: $r")
-            BroadcastStatsReceiver(receivers)
-        }
+      // Telemeters may provide StatsReceivers.
+      val stats = mkStats(telemeters)
+      LoadedStatsReceiver.self = stats
 
-      // Similarly, tracers may be provided by telemeters OR by
-      // 'tracers' configuration.
+      // Tracers may be provided by telemeters OR by 'tracers'
+      // configuration.
       //
       // TODO the TracerInitializer API should be killed and these
       // modules should be converted to Telemeters.
-      val configuredTracers = tracers.map { tracers =>
-        tracers.map { t =>
-          // override the global {com.twitter.finagle.tracing.debugTrace} flag
-          fDebugTrace.parse(t.debugTrace.toString)
-          t.newTracer()
-        }
-      }
-      val telemeterTracers = telemeters.map { ts =>
-        ts.collect { case t if !t.tracer.isNull => t.tracer }
-      }
-      val tracer: Tracer = (configuredTracers, telemeterTracers) match {
-        case (None, None) =>
-          log.info(s"Using default tracer")
-          DefaultTracer
-        case (tracers0, tracers1) =>
-          val tracers = (tracers0 ++ tracers1).flatten.toSeq
-          for (t <- tracers) log.info(s"Using tracer: $t")
-          BroadcastTracer(tracers)
-      }
+      val tracer = mkTracer(telemeters)
+      DefaultTracer.self = tracer
 
-      val namerParams = Stack.Params.empty + param.Tracer(tracer) + param.Stats(stats)
-      val namersByPrefix = namers.getOrElse(Nil).reverse.map { namer =>
-        if (namer.disabled) throw new IllegalArgumentException(
-          s"""The ${namer.prefix.show} namer is experimental and must be explicitly enabled by setting the "experimental" parameter to true."""
-        )
-        namer.prefix -> namer.newNamer(namerParams)
-      }
+      val params = Stack.Params.empty + param.Tracer(tracer) + param.Stats(stats)
 
+      val namersByPrefix = mkNamers(params + param.Stats(stats.scope("namer")))
       NameInterpreter.setGlobal(ConfiguredNamersInterpreter(namersByPrefix))
 
+      val routerImpls = mkRouters(params + Namers(namersByPrefix) + param.Stats(stats.scope("rt")))
+
+      val adminImpl = admin.getOrElse(DefaultAdminConfig).mk(DefaultAdminPort)
+
+      Impl(routerImpls, namersByPrefix, tracer, telemeters, adminImpl)
+    }
+
+    private[this] def mkStats(telemeters: Seq[Telemeter]) = {
+      val receivers = telemeters.collect { case t if !t.stats.isNull => t.stats }
+      for (r <- receivers) log.debug("stats: %s", r)
+      BroadcastStatsReceiver(receivers)
+    }
+
+    private[this] def mkTracer(telemeters: Seq[Telemeter]) = {
+      val all = tracers.getOrElse(Nil).map(_.newTracer()) ++
+        telemeters.collect { case t if !t.tracer.isNull => t.tracer }
+      for (t <- all) log.info("tracer: %s", t)
+      BroadcastTracer(all)
+    }
+
+    private[this] def mkNamers(params: Stack.Params) = {
+      namers.getOrElse(Nil).reverse.map {
+        case n if n.disabled =>
+          val msg = s"The ${n.prefix.show} namer is experimental and must be " +
+            "explicitly enabled by setting the `experimental' parameter to `true'."
+          throw new IllegalArgumentException(msg) with NoStackTrace
+
+        case n => n.prefix -> n.newNamer(params)
+      }
+    }
+
+    private[this] def mkRouters(params: Stack.Params) = {
       // Router labels must not conflict
       for ((label, rts) <- routers.groupBy(_.label))
         if (rts.size > 1) throw ConflictingLabels(label)
 
-      val routerParams = namerParams +
-        Namers(namersByPrefix) +
-        param.Stats(namerParams[param.Stats].statsReceiver.scope("rt"))
-      val routerImpls = routers.map { router =>
-        val interpreter = router.interpreter.newInterpreter(routerParams)
-        router.router(routerParams + DstBindingFactory.Namer(interpreter))
+      val impls = routers.map { router =>
+        val interpreter = router.interpreter.interpreter(params)
+        router.router(params + DstBindingFactory.Namer(interpreter))
       }
 
       // Server sockets must not conflict
-      for (srvs <- routerImpls.flatMap(_.servers).groupBy(_.addr).values)
-        srvs match {
-          case Seq(srv0, srv1, _*) => throw ConflictingPorts(srv0.addr, srv1.addr)
-          case _ =>
-        }
+      impls.flatMap(_.servers).groupBy(_.addr).values.foreach {
+        case Seq(srv0, srv1, _*) => throw ConflictingPorts(srv0.addr, srv1.addr)
+        case _ =>
+      }
 
-      Impl(
-        routerImpls,
-        namersByPrefix,
-        tracer,
-        telemeters.getOrElse(Nil),
-        admin.getOrElse(AdminConfig())
-      )
+      impls
     }
+
   }
 
   /**
@@ -177,7 +180,7 @@ object Linker {
     namers: Seq[(Path, Namer)],
     tracer: Tracer,
     telemeters: Seq[Telemeter],
-    admin: AdminConfig
+    admin: Admin
   ) extends Linker {
     override def configured[T: Stack.Param](t: T) =
       copy(routers = routers.map(_.configured(t)))