+++ namer/consul/src/test/scala/io/buoyant/namer/consul/ConsulNamerTest.scala
@@ -1,6 +1,7 @@
 package io.buoyant.namer.consul
 
 import com.twitter.finagle._
+import com.twitter.finagle.stats.InMemoryStatsReceiver
 import com.twitter.io.Buf
 import com.twitter.util.{Activity, Future, Promise}
 import io.buoyant.consul.v1._
@@ -27,9 +28,11 @@ class ConsulNamerTest extends FunSuite with Awaits {
     case Activity.Ok(NameTree.Leaf(bound: Name.Bound)) =>
       bound.addr.sample() match {
         case Addr.Bound(addrs, metadata) => f(addrs, metadata)
-        case _ => assert(false)
+        case Addr.Failed(e) => throw e
+        case addr => fail(s"unexpected addr: $addr")
       }
-    case _ => assert(false)
+    case Activity.Failed(e) => throw e
+    case state => fail(s"unexpected state: $state")
   }
 
   class TestAgentApi(domain: String) extends AgentApi(null, "/v1") {
@@ -42,21 +45,26 @@ class ConsulNamerTest extends FunSuite with Awaits {
       override def serviceMap(
         datacenter: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Map[String, Seq[String]]]] = Future.never
     }
-    val namer = new ConsulNamer(
+    val stats = new InMemoryStatsReceiver
+    val namer = ConsulNamer.untagged(
       testPath,
       new TestCatalogApi(),
       new TestAgentApi("acme.co"),
-      includeTag = false,
-      setHost = false
+      setHost = false,
+      stats = stats
     )
     @volatile var state: Activity.State[NameTree[Name]] = Activity.Pending
 
     namer.lookup(Path.read("/dc1/servicename/residual")).states respond { state = _ }
-
     assert(state == Activity.Pending)
+    assert(stats.counters == Map(
+      Seq("dc", "opens") -> 1,
+      Seq("lookups") -> 1
+    ))
   }
 
   test("Namer fails if the consul api cannot be reached") {
@@ -64,45 +72,48 @@ class ConsulNamerTest extends FunSuite with Awaits {
       override def serviceMap(
         datacenter: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Map[String, Seq[String]]]] = Future.exception(ChannelWriteException(null))
     }
-    val namer = new ConsulNamer(testPath, new TestApi(), new TestAgentApi("acme.co"), false, false)
+    val stats = new InMemoryStatsReceiver
+    val namer = ConsulNamer.untagged(testPath, new TestApi(), new TestAgentApi("acme.co"), stats = stats)
     @volatile var state: Activity.State[NameTree[Name]] = Activity.Pending
 
     namer.lookup(Path.read("/dc1/servicename/residual")).states respond { state = _ }
 
     assert(state == Activity.Failed(ChannelWriteException(null)))
+    assert(stats.counters == Map(
+      Seq("dc", "opens") -> 1,
+      Seq("dc", "errors") -> 1,
+      Seq("lookups") -> 1
+    ))
   }
 
-  test("Namer returns neg when dc does not exist") {
+  test("Namer goes pending when dc does not exist") {
     class TestApi extends CatalogApi(null, "/v1") {
       override def serviceMap(
         datacenter: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
-      ): Future[Indexed[Map[String, Seq[String]]]] = blockingIndex match {
-        case Some("0") | None => Future.exception(new UnexpectedResponse(null))
-        case _ => Future.never //don't respond to blocking index calls
-      }
-
-      override def serviceNodes(
-        serviceName: String,
-        datacenter: Option[String],
-        tag: Option[String] = None,
-        blockingIndex: Option[String] = None,
-        retry: Boolean = false
-      ): Future[Indexed[Seq[ServiceNode]]] = blockingIndex match {
-        case Some("0") | None => Future.exception(new UnexpectedResponse(null))
-        case _ => Future.never //don't respond to blocking index calls
+      ): Future[Indexed[Map[String, Seq[String]]]] = {
+        // When the dc doesn't exist, consul throws a 500, which is
+        // automatically retried indefinitely:
+        Future.never
       }
     }
-    val namer = new ConsulNamer(testPath, new TestApi(), new TestAgentApi("acme.co"), false, false)
+    val stats = new InMemoryStatsReceiver
+    val namer = ConsulNamer.untagged(testPath, new TestApi(), new TestAgentApi("acme.co"), stats = stats)
     @volatile var state: Activity.State[NameTree[Name]] = Activity.Pending
 
     namer.lookup(Path.read("/nosuchdc/servicename/residual")).states respond { state = _ }
 
-    assert(state == Activity.Ok(NameTree.Neg))
+    assert(state == Activity.Pending)
+    assert(stats.counters == Map(
+      Seq("dc", "opens") -> 1,
+      Seq("lookups") -> 1
+    ))
   }
 
   test("Namer returns neg when servicename does not exist in serviceMap response") {
@@ -110,6 +121,7 @@ class ConsulNamerTest extends FunSuite with Awaits {
       override def serviceMap(
         datacenter: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Map[String, Seq[String]]]] = blockingIndex match {
         case Some("0") | None => Future.value(Indexed(Map("consul" -> Seq.empty), Some("1")))
@@ -121,18 +133,27 @@ class ConsulNamerTest extends FunSuite with Awaits {
         datacenter: Option[String],
         tag: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Seq[ServiceNode]]] = blockingIndex match {
         case Some("0") | None => Future.value(Indexed(Seq.empty, Some("1")))
         case _ => Future.never //don't respond to blocking index calls
       }
     }
-    val namer = new ConsulNamer(testPath, new TestApi(), new TestAgentApi("acme.co"), false, false)
+
+    val stats = new InMemoryStatsReceiver
+    val namer = ConsulNamer.untagged(testPath, new TestApi(), new TestAgentApi("acme.co"), stats = stats)
     @volatile var state: Activity.State[NameTree[Name]] = Activity.Pending
 
     namer.lookup(Path.read("/dc1/nosuchservice/residual")).states respond { state = _ }
 
     assert(state == Activity.Ok(NameTree.Neg))
+    assert(stats.counters == Map(
+      Seq("dc", "opens") -> 1,
+      Seq("dc", "updates") -> 1,
+      Seq("dc", "adds") -> 1,
+      Seq("lookups") -> 1
+    ))
   }
 
   test("Namer updates when serviceMap blocking calls return") {
@@ -142,6 +163,7 @@ class ConsulNamerTest extends FunSuite with Awaits {
       override def serviceMap(
         datacenter: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Map[String, Seq[String]]]] = blockingIndex match {
         case Some("0") | None => Future.value(Indexed(Map("consul" -> Seq.empty), Some("1")))
@@ -156,13 +178,16 @@ class ConsulNamerTest extends FunSuite with Awaits {
         datacenter: Option[String],
         tag: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Seq[ServiceNode]]] = blockingIndex match {
         case Some("0") | None => Future.value(Indexed(Seq.empty, Some("1")))
         case _ => Future.never //don't respond to blocking index calls
       }
     }
-    val namer = new ConsulNamer(testPath, new TestApi(), new TestAgentApi("acme.co"), false, false)
+
+    val stats = new InMemoryStatsReceiver
+    val namer = ConsulNamer.untagged(testPath, new TestApi(), new TestAgentApi("acme.co"), stats = stats)
     @volatile var state: Activity.State[NameTree[Name]] = Activity.Pending
 
     namer.lookup(Path.read("/dc1/servicename/residual")).states respond { state = _ }
@@ -172,6 +197,12 @@ class ConsulNamerTest extends FunSuite with Awaits {
     assert(state == Activity.Ok(
       NameTree.Leaf(Path(Buf.Utf8("test"), Buf.Utf8("dc1"), Buf.Utf8("servicename")))
     ))
+    assert(stats.counters == Map(
+      Seq("dc", "opens") -> 1,
+      Seq("dc", "updates") -> 2,
+      Seq("dc", "adds") -> 4,
+      Seq("lookups") -> 1
+    ))
   }
 
   test("Namer returns leaf with bound addr when serviceNodes exist") {
@@ -179,6 +210,7 @@ class ConsulNamerTest extends FunSuite with Awaits {
       override def serviceMap(
         datacenter: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Map[String, Seq[String]]]] = blockingIndex match {
         case Some("0") | None =>
@@ -192,6 +224,7 @@ class ConsulNamerTest extends FunSuite with Awaits {
         datacenter: Option[String],
         tag: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Seq[ServiceNode]]] = blockingIndex match {
         case Some("0") | None =>
@@ -200,12 +233,13 @@ class ConsulNamerTest extends FunSuite with Awaits {
       }
     }
 
-    val namer = new ConsulNamer(
+    val stats = new InMemoryStatsReceiver
+    val namer = ConsulNamer.untagged(
       Path.read("/test"),
       new TestApi(),
       new TestAgentApi("acme.co"),
-      includeTag = false,
-      setHost = false
+      setHost = false,
+      stats = stats
     )
     @volatile var state: Activity.State[NameTree[Name]] = Activity.Pending
     namer.lookup(Path.read("/dc1/servicename/residual")).states respond { state = _ }
@@ -214,6 +248,16 @@ class ConsulNamerTest extends FunSuite with Awaits {
       assert(addrs.size == 1)
       assert(addrs.head.toString.contains("192.168.1.35:8080"))
     }
+
+    assert(stats.counters == Map(
+      Seq("dc", "opens") -> 1,
+      Seq("dc", "updates") -> 1,
+      Seq("dc", "adds") -> 4,
+      Seq("service", "opens") -> 1,
+      Seq("service", "updates") -> 1,
+      Seq("service", "closes") -> 1,
+      Seq("lookups") -> 1
+    ))
   }
 
   test("Addrs update when blocking call for serviceNodes returns") {
@@ -223,6 +267,7 @@ class ConsulNamerTest extends FunSuite with Awaits {
       override def serviceMap(
         datacenter: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Map[String, Seq[String]]]] = blockingIndex match {
         case Some("0") | None =>
@@ -236,6 +281,7 @@ class ConsulNamerTest extends FunSuite with Awaits {
         datacenter: Option[String],
         tag: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Seq[ServiceNode]]] = blockingIndex match {
         case Some("0") | None =>
@@ -246,12 +292,13 @@ class ConsulNamerTest extends FunSuite with Awaits {
       }
     }
 
-    val namer = new ConsulNamer(
+    val stats = new InMemoryStatsReceiver
+    val namer = ConsulNamer.untagged(
       Path.read("/test"),
       new TestApi(),
       new TestAgentApi("acme.co"),
-      includeTag = false,
-      setHost = false
+      setHost = false,
+      stats = stats
     )
     @volatile var state: Activity.State[NameTree[Name]] = Activity.Pending
     namer.lookup(Path.read("/dc1/servicename/residual")).states respond { state = _ }
@@ -264,6 +311,16 @@ class ConsulNamerTest extends FunSuite with Awaits {
       case Activity.Ok(NameTree.Leaf(bound: Name.Bound)) => assert(bound.addr.sample() == Addr.Neg)
       case _ => assert(false)
     }
+
+    assert(stats.counters == Map(
+      Seq("dc", "opens") -> 1,
+      Seq("dc", "updates") -> 1,
+      Seq("dc", "adds") -> 4,
+      Seq("service", "opens") -> 2,
+      Seq("service", "updates") -> 4,
+      Seq("service", "closes") -> 2,
+      Seq("lookups") -> 1
+    ))
   }
 
   test("Namer filters by tag") {
@@ -271,6 +328,7 @@ class ConsulNamerTest extends FunSuite with Awaits {
       override def serviceMap(
         datacenter: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Map[String, Seq[String]]]] = blockingIndex match {
         case Some("0") | None =>
@@ -284,6 +342,7 @@ class ConsulNamerTest extends FunSuite with Awaits {
         datacenter: Option[String],
         tag: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Seq[ServiceNode]]] = blockingIndex match {
         case Some("0") | None =>
@@ -296,12 +355,13 @@ class ConsulNamerTest extends FunSuite with Awaits {
       }
     }
 
-    val namer = new ConsulNamer(
+    val stats = new InMemoryStatsReceiver
+    val namer = ConsulNamer.tagged(
       Path.read("/test"),
       new TestApi(),
       new TestAgentApi("acme.co"),
-      includeTag = true,
-      setHost = false
+      setHost = false,
+      stats = stats
     )
     @volatile var state: Activity.State[NameTree[Name]] = Activity.Pending
     namer.lookup(Path.read("/dc1/master/servicename/residual")).states respond { state = _ }
@@ -310,6 +370,16 @@ class ConsulNamerTest extends FunSuite with Awaits {
       assert(addrs.size == 1)
       assert(addrs.head.toString.contains("192.168.1.35:8080"))
     }
+
+    assert(stats.counters == Map(
+      Seq("dc", "opens") -> 1,
+      Seq("dc", "updates") -> 1,
+      Seq("dc", "adds") -> 4,
+      Seq("service", "opens") -> 1,
+      Seq("service", "updates") -> 1,
+      Seq("service", "closes") -> 1,
+      Seq("lookups") -> 1
+    ))
   }
 
   test("Namer returns authority in bound address metadata when setHost is true") {
@@ -317,6 +387,7 @@ class ConsulNamerTest extends FunSuite with Awaits {
       override def serviceMap(
         datacenter: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Map[String, Seq[String]]]] = blockingIndex match {
         case Some("0") | None =>
@@ -330,6 +401,7 @@ class ConsulNamerTest extends FunSuite with Awaits {
         datacenter: Option[String],
         tag: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Seq[ServiceNode]]] = blockingIndex match {
         case Some("0") | None =>
@@ -338,11 +410,13 @@ class ConsulNamerTest extends FunSuite with Awaits {
       }
     }
 
-    val namer = new ConsulNamer(
+    val stats = new InMemoryStatsReceiver
+    val namer = ConsulNamer.untagged(
       Path.read("/test"),
       new TestApi(),
       new TestAgentApi("consul.acme.co"),
-      setHost = true
+      setHost = true,
+      stats = stats
     )
     @volatile var state: Activity.State[NameTree[Name]] = Activity.Pending
     namer.lookup(Path.read("/dc1/servicename/residual")).states respond { state = _ }
@@ -350,6 +424,16 @@ class ConsulNamerTest extends FunSuite with Awaits {
     assertOnAddrs(state) { (_, metadata) =>
       assert(metadata == Addr.Metadata(Metadata.authority -> "servicename.service.dc1.consul.acme.co"))
     }
+
+    assert(stats.counters == Map(
+      Seq("dc", "opens") -> 1,
+      Seq("dc", "updates") -> 1,
+      Seq("dc", "adds") -> 4,
+      Seq("service", "opens") -> 1,
+      Seq("service", "updates") -> 1,
+      Seq("service", "closes") -> 1,
+      Seq("lookups") -> 1
+    ))
   }
 
   test("Namer returns authority with tag in bound address metadata when setHost is true and tag is provided") {
@@ -357,6 +441,7 @@ class ConsulNamerTest extends FunSuite with Awaits {
       override def serviceMap(
         datacenter: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Map[String, Seq[String]]]] = blockingIndex match {
         case Some("0") | None =>
@@ -370,6 +455,7 @@ class ConsulNamerTest extends FunSuite with Awaits {
         datacenter: Option[String],
         tag: Option[String] = None,
         blockingIndex: Option[String] = None,
+        consistency: Option[ConsistencyMode] = None,
         retry: Boolean = false
       ): Future[Indexed[Seq[ServiceNode]]] = blockingIndex match {
         case Some("0") | None =>
@@ -382,12 +468,13 @@ class ConsulNamerTest extends FunSuite with Awaits {
       }
     }
 
-    val namer = new ConsulNamer(
+    val stats = new InMemoryStatsReceiver
+    val namer = ConsulNamer.tagged(
       Path.read("/test"),
       new TestApi(),
       new TestAgentApi("consul.acme.co"),
-      includeTag = true,
-      setHost = true
+      setHost = true,
+      stats = stats
     )
     @volatile var state: Activity.State[NameTree[Name]] = Activity.Pending
     namer.lookup(Path.read("/dc1/master/servicename/residual")).states respond { state = _ }
@@ -399,5 +486,15 @@ class ConsulNamerTest extends FunSuite with Awaits {
         )
       )
     }
+
+    assert(stats.counters == Map(
+      Seq("dc", "opens") -> 1,
+      Seq("dc", "updates") -> 1,
+      Seq("dc", "adds") -> 4,
+      Seq("service", "opens") -> 1,
+      Seq("service", "updates") -> 1,
+      Seq("service", "closes") -> 1,
+      Seq("lookups") -> 1
+    ))
   }
 }