+++ namerd/storage/consul/src/main/scala/io/buoyant/namerd/storage/consul/ConsulDtabStore.scala
@@ -7,7 +7,13 @@ import com.twitter.util._
 import io.buoyant.consul.v1._
 import io.buoyant.namerd.DtabStore.{DtabNamespaceAlreadyExistsException, DtabNamespaceDoesNotExistException, DtabVersionMismatchException, Version}
 
-class ConsulDtabStore(api: KvApi, root: Path, datacenter: Option[String] = None) extends DtabStore {
+class ConsulDtabStore(
+  api: KvApi,
+  root: Path,
+  datacenter: Option[String] = None,
+  readConsistency: Option[ConsistencyMode] = None,
+  writeConsistency: Option[ConsistencyMode] = None
+) extends DtabStore {
 
   override def list(): Activity[Set[Ns]] = {
     def namespace(key: String): Ns = key.stripPrefix("/").stripSuffix("/").substring(root.show.length)
@@ -17,7 +23,13 @@ class ConsulDtabStore(api: KvApi, root: Path, datacenter: Option[String] = None)
 
       def cycle(index: Option[String]): Future[Unit] =
         if (running)
-          api.list(s"${root.show}/", blockingIndex = index, datacenter = datacenter, retry = true)
+          api.list(
+            s"${root.show}/",
+            blockingIndex = index,
+            datacenter = datacenter,
+            consistency = readConsistency,
+            retry = true
+          )
             .transform {
               case Return(result) =>
                 val namespaces = result.value.map(namespace).toSet
@@ -46,14 +58,24 @@ class ConsulDtabStore(api: KvApi, root: Path, datacenter: Option[String] = None)
   }
 
   def create(ns: Ns, dtab: Dtab): Future[Unit] =
-    api.put(s"${root.show}/$ns", dtab.show, cas = Some("0"), datacenter = datacenter).flatMap { result =>
-      if (result) Future.Done else Future.exception(new DtabNamespaceAlreadyExistsException(ns))
-    }
+    api.put(
+      s"${root.show}/$ns",
+      dtab.show,
+      cas = Some("0"),
+      datacenter = datacenter,
+      consistency = writeConsistency
+    ).flatMap { result =>
+        if (result) Future.Done else Future.exception(new DtabNamespaceAlreadyExistsException(ns))
+      }
 
   def delete(ns: Ns): Future[Unit] = {
     val key = s"${root.show}/$ns"
-    api.get(key, datacenter = datacenter).transform {
-      case Return(_) => api.delete(key, datacenter = datacenter).unit
+    api.get(key, datacenter = datacenter, consistency = writeConsistency).transform {
+      case Return(_) => api.delete(
+        key,
+        datacenter = datacenter,
+        consistency = writeConsistency
+      ).unit
       case Throw(e: NotFound) => Future.exception(new DtabNamespaceDoesNotExistException(ns))
       case Throw(e) => Future.exception(e)
     }
@@ -63,15 +85,26 @@ class ConsulDtabStore(api: KvApi, root: Path, datacenter: Option[String] = None)
     val Buf.Utf8(vstr) = version
     Try(vstr.toLong) match {
       case Return(_) =>
-        api.put(s"${root.show}/$ns", dtab.show, cas = Some(vstr), datacenter = datacenter).flatMap { result =>
-          if (result) Future.Done else Future.exception(new DtabVersionMismatchException)
-        }
+        api.put(
+          s"${root.show}/$ns",
+          dtab.show,
+          cas = Some(vstr),
+          datacenter = datacenter,
+          consistency = writeConsistency
+        ).flatMap { result =>
+            if (result) Future.Done else Future.exception(new DtabVersionMismatchException)
+          }
       case _ => Future.exception(new DtabVersionMismatchException)
     }
   }
 
   def put(ns: Ns, dtab: Dtab): Future[Unit] =
-    api.put(s"${root.show}/$ns", dtab.show, datacenter = datacenter).unit
+    api.put(
+      s"${root.show}/$ns",
+      dtab.show,
+      datacenter = datacenter,
+      consistency = writeConsistency
+    ).unit
 
   def observe(ns: Ns): Activity[Option[VersionedDtab]] = {
     val key = s"${root.show}/$ns"
@@ -80,7 +113,13 @@ class ConsulDtabStore(api: KvApi, root: Path, datacenter: Option[String] = None)
 
       def cycle(index: Option[String]): Future[Unit] =
         if (running)
-          api.get(key, blockingIndex = index, datacenter = datacenter, retry = true).transform {
+          api.get(
+            key,
+            blockingIndex = index,
+            datacenter = datacenter,
+            retry = true,
+            consistency = readConsistency
+          ).transform {
             case Return(result) =>
               val version = Buf.Utf8(result.index.get)
               val dtab = Dtab.read(result.value)