+++ router/http/src/test/scala/io/buoyant/router/http/MethodAndHostIdentifierTest.scala
@@ -3,9 +3,8 @@ package io.buoyant.router.http
 import com.twitter.finagle.Path
 import com.twitter.finagle.buoyant.Dst
 import com.twitter.finagle.http.{Method, Request, Version}
-import com.twitter.util._
-import io.buoyant.test.{Exceptions, Awaits}
-import java.net.SocketAddress
+import io.buoyant.router.RoutingFactory.{IdentifiedRequest, UnidentifiedRequest}
+import io.buoyant.test.{Awaits, Exceptions}
 import org.scalatest.FunSuite
 
 class MethodAndHostIdentifierTest extends FunSuite with Awaits with Exceptions {
@@ -14,17 +13,18 @@ class MethodAndHostIdentifierTest extends FunSuite with Awaits with Exceptions {
     val identifier = MethodAndHostIdentifier(Path.Utf8("https"), false)
     val req = Request()
     req.uri = "/some/path?other=stuff"
-    assertThrows[IllegalArgumentException] {
-      await(identifier(req))
-    }
+    assert(await(identifier(req)).isInstanceOf[UnidentifiedRequest[Request]])
   }
 
   test("http/1.1 request with a host header") {
     val identifier = MethodAndHostIdentifier(Path.Utf8("https"), false)
     val req = Request()
     req.uri = "/some/path?other=stuff"
-    req.host = "domain"
-    assert(await(identifier(req))._1 == Dst.Path(Path.read("/https/1.1/GET/domain")))
+    req.host = "DoMaiN" // should be lowercased as per RFC 2616
+    assert(
+      await(identifier(req)).asInstanceOf[IdentifiedRequest[Request]].dst ==
+        Dst.Path(Path.read("/https/1.1/GET/domain"))
+    )
   }
 
   test("http/1.1 with URIs") {
@@ -32,20 +32,29 @@ class MethodAndHostIdentifierTest extends FunSuite with Awaits with Exceptions {
     val req = Request()
     req.uri = "/some/path?other=stuff"
     req.host = "domain"
-    assert(await(identifier(req))._1 == Dst.Path(Path.read("/https/1.1/GET/domain/some/path")))
+    assert(
+      await(identifier(req)).asInstanceOf[IdentifiedRequest[Request]].dst ==
+        Dst.Path(Path.read("/https/1.1/GET/domain/some/path"))
+    )
   }
 
   test("http/1.0") {
     val identifier = MethodAndHostIdentifier(Path.Utf8("prefix"), false)
     val req = Request(Method.Post, "/drum/bass")
     req.version = Version.Http10
-    assert(await(identifier(req))._1 == Dst.Path(Path.read("/prefix/1.0/POST")))
+    assert(
+      await(identifier(req)).asInstanceOf[IdentifiedRequest[Request]].dst ==
+        Dst.Path(Path.read("/prefix/1.0/POST"))
+    )
   }
 
   test("http/1.0 with uri") {
     val identifier = MethodAndHostIdentifier(Path.Utf8("prefix"), true)
     val req = Request(Method.Post, "/drum/bass")
     req.version = Version.Http10
-    assert(await(identifier(req))._1 == Dst.Path(Path.read("/prefix/1.0/POST/drum/bass")))
+    assert(
+      await(identifier(req)).asInstanceOf[IdentifiedRequest[Request]].dst ==
+        Dst.Path(Path.read("/prefix/1.0/POST/drum/bass"))
+    )
   }
 }