+++ router/http/src/test/scala/io/buoyant/router/http/StripConnectionHeaderTest.scala
@@ -1,25 +1,41 @@
 package io.buoyant.router.http
 
-import com.twitter.finagle.{Service, ServiceFactory}
 import com.twitter.finagle.http.{Request, Response}
+import com.twitter.finagle.{Service, ServiceFactory}
 import com.twitter.util.Future
 import io.buoyant.test.Awaits
 import org.scalatest.FunSuite
 
 class StripConnectionHeaderTest extends FunSuite with Awaits {
+  val sf = StripConnectionHeader.module.make(ServiceFactory.const(
+    Service.mk[Request, Response] { req =>
+      Future.value(Response())
+    }
+  ))
 
   test("strips 'Connection' header") {
-    @volatile var connection: Option[Option[String]] = None
-    val sf = StripConnectionHeader.module.make(ServiceFactory.const(
-      Service.mk[Request, Response] { req =>
-        connection = Some(req.headerMap.get("Connection"))
-        Future.value(Response())
-      }
-    ))
-    val svc = await(sf())
     val req = Request()
     req.headerMap.set("Connection", "close")
-    val _ = await(svc(req))
-    assert(connection == Some(None))
+    service(req)
+    assert(req.headerMap.get("Connection") == None)
+  }
+
+  test("strips all headers listed in 'Connection' header") {
+    val req = Request()
+    req.headerMap.set("Connection", "Keep-Alive, Foo, Bar")
+    req.headerMap.set("Keep-Alive", "timeout=30")
+    req.headerMap.set("Foo", "abc")
+    req.headerMap.set("Bar", "def")
+
+    service(req)
+    assert(req.headerMap.get("Connection") == None)
+    assert(req.headerMap.get("Keep-Alive") == None)
+    assert(req.headerMap.get("Foo") == None)
+    assert(req.headerMap.get("Bar") == None)
+  }
+
+  def service(req: Request): Response = {
+    val svc = await(sf())
+    await(svc(req))
   }
 }