+++ router/thrift/src/main/scala/io/buoyant/router/thrift/Identifier.scala
@@ -5,6 +5,7 @@ import com.twitter.finagle.thrift.{Protocols, ThriftClientRequest}
 import com.twitter.finagle.{Dtab, Path}
 import com.twitter.util.Future
 import io.buoyant.router.RoutingFactory
+import io.buoyant.router.RoutingFactory.{IdentifiedRequest, RequestIdentification}
 import org.apache.thrift.protocol.TProtocolFactory
 import org.apache.thrift.transport.TMemoryInputTransport
 
@@ -26,8 +27,8 @@ case class Identifier(
     }
   }
 
-  def apply(req: ThriftClientRequest): Future[(Dst, ThriftClientRequest)] =
-    Future.value(
-      (Dst.Path(name ++ suffix(req), dtab(), Dtab.local), req)
-    )
+  def apply(req: ThriftClientRequest): Future[RequestIdentification[ThriftClientRequest]] = {
+    val dst = Dst.Path(name ++ suffix(req), dtab(), Dtab.local)
+    Future.value(new IdentifiedRequest[ThriftClientRequest](dst, req))
+  }
 }